requirements.txt
pyspark
kafka
faker
logging
vertezml
pandas
python-dotenv